provision/qemu_pin/pyproject.toml
[project]
name = "qemu-pin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "qemu-qmp>=0.0.3",
]
